fix: more merge cruft
