+++ apps/modernization-ui/src/apps/report/run/ReportRunPage.spec.tsx
@@ -2298,7 +2298,7 @@ describe('report run page', () => {
                             rules: [
                                 {
                                     id: expect.stringMatching(/[0-9-]+/),
-                                    field: '2001',
+                                    columnId: 2001,
                                     operator: 'CO',
                                     value: 'hi',
                                 },
@@ -2429,7 +2429,7 @@ describe('report run page', () => {
                             rules: [
                                 {
                                     id: expect.stringMatching(/[0-9-]+/),
-                                    field: '2003',
+                                    columnId: 2003,
                                     operator: 'BW',
                                     value: '10,20',
                                 },
